explicit abort: the planner withdraws a trajectory instead of inventing a manoeuvre
Decision from review: option 5. The planner has no basis for choosing a
recovery for a given airframe, and every manoeuvre it could synthesise
risks another unflyable trajectory — the one stop it can build is a
zero-velocity hold and this platform's cruise minimum is 122 m/s. So it
declares the flight withdrawn and the execution layer, which knows the
vehicle, owns what happens next.

WIRE. PolyTraj gains trajectory_id (uint64, monotone per node).
LocalTrajData::traj_id already existed and simply never reached the wire —
and a field by that name was removed in the 2026-07 audit for being
written and never read, so the message now records that something reads
it, to stop the next audit removing it again. start_time cannot stand in
for identity: it is a clock reading and moves under replay and time
correction.

TrajectoryExecutionControl carries the id, ACTION_ABORT, a reason that
distinguishes "it is blocked" from "we can no longer judge it", and
diagnostics. It publishes on /planning/execution_control with the same
RELIABLE + TRANSIENT_LOCAL QoS as the trajectory, so a late consumer
receives both and discards rather than starts.

ORDER. The abort goes out BEFORE the local slot is cleared — otherwise
the FSM would announce an id it can no longer read. Then the slot is
invalidated, then the FSM moves to WAIT_EXTERNAL_RECOVERY, a state that
issues nothing. EMERGENCY_STOP is deliberately unused.

Detection stays in PathManager (it owns the SDF and the stored
trajectory), announcement in ReplanFSM (it owns the publishers and the
state machine), joined by a hook — so a PathManager-internal trigger like
the deferred-obstacle flush reaches the wire too.

Zones withdraw only when the set actually CHANGES, compared by content
fingerprint. A periodic republish of the same zones changes nothing about
the flight, and grounding on it would make the fail-closed rule
unusable.

REGRESSIONS. exec_abort_test is new and drives real publishers and
subscribers at production QoS against a consumer implementation: a
trajectory is executed, an abort naming it stops the follower, a
duplicate is idempotent, a trajectory issued after the abort still flies,
a replayed abort for an older id does NOT stop the current one, and a
late joiner does not start flying an aborted latched trajectory.
envchange pins the producer half — nothing announced on a harmless
change, exactly one announcement on a block, carrying the id of the
trajectory being withdrawn and the reason. Mutations: moving the hook
after the slot clear loses the id; announcing on every check kills the
three silence assertions.

Found on the way: state_str was sized 6 while FSM_EXEC_STATE grew to 7,
so the last state indexed past the end. Sized from the enum with a
static_assert.

NOT closed, and stated: no follower implementation exists in this repo,
so the test implements the consumer contract itself. That verifies the
contract is coherent, not that an external follower obeys it — that needs
agreement on the follower side. ACK status is not published (the contract
called it optional).

69/69 chain variants, exec_abort_test, risk harness 154/0, start_claim /
transition / waypoint / terrain_risk_mask all pass.

+++ src/path_optimizer/include/path_optimizer/plan_container.hpp
@@ -31,7 +31,11 @@ namespace ego_planner
   {
     poly_traj::Trajectory traj;
     int drone_id; // A negative value indicates no received trajectories.
-    int traj_id;
+    // Monotone per node, never reset while a mission runs. On the wire as
+    // PolyTraj.trajectory_id, and TrajectoryExecutionControl names it to say
+    // WHICH trajectory an abort refers to — so it must not wrap during a
+    // flight, hence 64 bits rather than the int it was while nothing read it.
+    uint64_t traj_id;
     double duration;
     double start_time; // world time
     double end_time;   // world time
